feat(test): Add test DB options def=sqlite #227
Sometimes it is not convenient to run tests if a connection to a
PostgreSQL, MySQL or other DB is unavailable.
Local and or learning projects may not warrant a Production
grade Database.
Using an ENV to select between db.sqlite3 or a production database can
make the testing process easier.

closes #227