chore(settings): Update default Allowed hosts config #636 (#649)
* chore(settings): Update default Allowed hosts config #636

Previously `ALLOWED_HOSTS` and `INTERNAL_IPS` would be configured with
the options passed in with cookiecutter.  The settings files have all
been updated to take their values either from `.env` or `ENV VARS`.
The `LOCAL` settings file has been modified to take a default
`ALLOWED_HOSTS` of `127.0.0.1` if `ALLOWED_HOSTS` is empty.  This makes
setup easy for new django users and will get them up and running
quickly.

All other settings are taking the values supplied in `ALLOWED_HOSTS` as
the default, and the primary source is either `.env` or `ENV_VARS`.

closes #636

* test(settings): Update and add tests for env #636

Changes to the environment settings required some tweaks to existing
tests, and inclusion of new tests.

closes #636